requirements.txt
pandas==2.2.2
fastapi==0.111.0
requests==2.32.3
psycopg2==2.9.9
asyncio==3.4.3
faker==33.1.0
scikit-learn==1.3.2
numpy<2.0.0
uvicorn==0.30.1
gunicorn==23.0.0
tensorflow==2.16.1
matplotlib==3.9.3
plotly==5.24.1
seaborn==0.13.2
xgboost==2.1.3
catboost==1.2.7
dash==2.18.2
python-dotenv==1.0.1